feat(ELE-2439): deprecate PollingAnalyzer + promote analytics_reports (D7 part 1)
## PollingAnalyzer deprecation
- Construction now emits DeprecationWarning
- Docstring includes migration guide:
  - Survey pipeline: siege_utilities.survey (build_chain + chain_to_argument)
  - Stats primitives: siege_utilities.data.statistics
  - Longitudinal / wave comparisons: survey.WaveSet (ELE-2440, coming)
- References ADR 0006 for rationale
- Class and module stay in place for one deprecation window; full
  removal deferred to next minor release (pairs with ELE-2440 landing
  the wave subsystem replacement)

## analytics_reports.py promotion
- Moved: reporting/analytics/analytics_reports.py -> reporting/analytics_reports.py
- Old path kept as a deprecation shim re-exporting from the new location
- reporting/__init__.py lazy-import table updated

## Deferred to ELE-2437 landing
- Extract longitudinal + change-detection to data/statistics/longitudinal.py
  (gates on data/statistics/ existing)
- Full removal of reporting/analytics/ subdirectory (gates on PollingAnalyzer
  deletion next release)

## Tests
- tests/test_polling_analyzer_deprecation.py (4 tests) covering:
  - Construction emits DeprecationWarning with migration text
  - New analytics_reports path imports
  - Old path shim fires DeprecationWarning
  - Top-level lazy export still resolves
- 225 pre-existing polling_analyzer + backward_compat tests pass
